Alterações finais

+++ mvc/dao/mysql/UsuarioDAO.php
@@ -20,11 +20,23 @@ class UsuarioDAO extends MysqlFactory implements IUsuarioDAO{
             ":email" => $email, 
             ":senha" => $senha
         ];
-        $retorno = $this->banco->executar($sql, $param);
-        return $retorno;
+        $this->banco->executar($sql, $param);
+        return $param;
     }
 
     public function alterar($id, $nome, $email, $senha){
+        $sqlVerifica = "SELECT id FROM usuarios WHERE id = :id";
+        $paramVerifica = [":id" => $id];
+        $usuarioExiste = $this->banco->executar($sqlVerifica, $paramVerifica);
+        if (count($usuarioExiste) == 0) {
+            $msg = array('erro' => "Usuario não existe!");
+            $jmsg = json_encode($msg);
+            http_response_code(400);
+            header("Content-Type: application/json");
+            echo $jmsg;
+            exit;
+        }
+
         $sql = "update usuarios set nome = :nome, email = :email, senha = :senha where id = :id";
         $param = [
             ":id" => $id, 
@@ -32,24 +44,46 @@ class UsuarioDAO extends MysqlFactory implements IUsuarioDAO{
             ":email" => $email, 
             ":senha" => $senha
         ];
-        $retorno = $this->banco->executar($sql, $param);
-        return $retorno;
+        $this->banco->executar($sql, $param);
+        return $param;
     }
 
     public function excluir($id){
-    // Exclui todas as inscrições do usuário
-    $sqlInscricoes = "DELETE FROM inscricoes WHERE usuario_id = :id";
-    $param = [":id" => $id];
-    $this->banco->executar($sqlInscricoes, $param);
+        $sqlVerifica = "SELECT id FROM usuarios WHERE id = :id";
+        $paramVerifica = [":id" => $id];
+        $usuarioExiste = $this->banco->executar($sqlVerifica, $paramVerifica);
+        if (count($usuarioExiste) == 0) {
+            $msg = array('erro' => "Usuario não existe!");
+            $jmsg = json_encode($msg);
+            http_response_code(400);
+            header("Content-Type: application/json");
+            echo $jmsg;
+            exit;
+        }
 
-    // Agora exclui o usuário
-    $sqlUsuario = "DELETE FROM usuarios WHERE id = :id";
-    $this->banco->executar($sqlUsuario, $param);
-
-    return true;
+        // Exclui todas as inscrições do usuário
+        $sqlInscricoes = "DELETE FROM inscricoes WHERE usuario_id = :id";
+        $param = [":id" => $id];
+        $this->banco->executar($sqlInscricoes, $param);
+        // Agora exclui o usuário
+        $sqlUsuario = "DELETE FROM usuarios WHERE id = :id";
+        $this->banco->executar($sqlUsuario, $param);
+        return $param;
     }
 
     public function autenticar($email, $senha){
+        $sqlVerifica = "SELECT email,senha FROM usuarios WHERE email = :email AND senha = :senha";
+        $paramVerifica = [":email" => $email, ":senha" => $senha];
+        $usuarioExiste = $this->banco->executar($sqlVerifica, $paramVerifica);
+        if (count($usuarioExiste) == 0) {
+            $msg = array('erro' => "Email ou senha incorretos!");
+            $jmsg = json_encode($msg);
+            http_response_code(400);
+            header("Content-Type: application/json");
+            echo $jmsg;
+            exit;
+        }
+
         $sql = "select id, nome, senha, email from usuarios where email = :email and senha = :senha";
         $param = [
             ":email" => $email,
@@ -58,36 +92,4 @@ class UsuarioDAO extends MysqlFactory implements IUsuarioDAO{
         $retorno = $this->banco->executar($sql, $param);
         return $retorno;
     }
-
-    /*public function getID($email){
-    $sql = "select id from usuarios where email = :email";
-    $param = [
-        ":email" => $email
-    ];
-    $retorno = $this->banco->executar($sql, $param);
-    return $retorno;
-    }*/
-
-    /*public function login(){
-        header("login.php");
-    }*/
-
-    /*public function listarId($id){
-        $sql = "select id,nome,senha,email from usuarios where id =:id";
-        $param = [
-            ":id" => $id
-        ];
-        $retorno = $this->banco->executar($sql, $param);
-        return $retorno;
-    }*/
-
-    /*public function fazerLogin($email,$senha){
-        $sql = "select id,nome,senha,email from usuarios where email = :email and senha = :senha";
-        $param = [
-            ":email" => $email,
-            ":senha" => $senha
-        ];
-        $retorno = $this->banco->executar($sql, $param);
-        return $retorno;
-    }*/
 }
